pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acji"
version = "0.1.0"
description = "Austin Coffee & Juice Intelligence ETL and analytics project"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "ACJI Team" }
]
dependencies = [
  "apscheduler>=3.10.4",
  "duckdb>=1.1.3",
  "dbt-duckdb>=1.8.2",
  "great_expectations==0.18.21",
  "joblib>=1.4.2",
  "numpy>=1.26.4",
  "pandas>=2.2.3",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "scikit-learn>=1.5.2",
  "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
